+++ database/migrations/2022_06_14_134606_create_results_table.php
@@ -15,8 +15,8 @@ class CreateResultsTable extends Migration
     {
         Schema::create('results', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id');
-            $table->integer('question_id');
+            $table->bigInteger('user_id');
+            $table->bigInteger('question_id');
             $table->timestamps();
         });
     }
